Refactor fluid simulation constants and presets
