+++ simulations/fluid-simulation/assets/js/fluid-simulation-logic.js
@@ -1,239 +1,7 @@
 // Import lil-gui
 import GUI from 'https://cdn.jsdelivr.net/npm/lil-gui@0.18/dist/lil-gui.esm.min.js';
-
-// --- Constants ---
-const FRUSTUM_SIZE = 120;
-const MAX_PARTICLES = 20000; // Max particles buffer can hold
-const DEFAULT_PRESET_NAME = 'Medium';
-
-// --- Presets Definition ---
-const PRESETS = {
-    'Low': {
-        particlesCount: 2000,
-        damping: 0.97,
-        maxVelocity: 2.0,
-        particleRepulsionRadius: 2.5,
-        repulsionStrength: 0.15,
-        cohesionRadius: 5.0,
-        cohesionStrength: 0.005,
-        particleSize: 1.5,
-        bloomEnabled: false,
-        enableObstacles: false,
-        edgeDamping: 0.6, containerRadius: 55, gravityStrength: 0.04, gravityEnabled: false,
-        interactionRadius: 10.0, attractionStrength: 0.15, repellingStrength: 0.30, vortexStrength: 0.15, stirStrength: 0.10, showInteractionRadius: true,
-        particleBaseColor: '#40E0D0', particleVelocityColorScale: 3.0, bgColor: '#050510',
-        bloomStrength: 0.4, bloomRadius: 0.3, bloomThreshold: 0.65,
-        wrapScreenEdges: false,
-        obstacles: [ { x: -25, y: 20, radius: 10 }, { x: 30, y: -15, radius: 14 }, { x: 0, y: -35, radius: 6 }],
-        emitters: [],
-        forceZones: [],
-        mouseTool: 'default',
-    },
-    'Medium': { // Base/default
-        particlesCount: 5000,
-        damping: 0.96,
-        maxVelocity: 2.5,
-        particleRepulsionRadius: 3.5,
-        repulsionStrength: 0.25,
-        cohesionRadius: 7.0,
-        cohesionStrength: 0.01,
-        particleSize: 1.8,
-        bloomEnabled: true,
-        bloomStrength: 0.4,
-        bloomRadius: 0.3,
-        bloomThreshold: 0.65,
-        wrapAroundEdges: false,
-        wrapScreenEdges: false,
-        useVerticalGradient: false,
-        gradientHueScale: 0.0,
-        enableObstacles: true,
-        edgeDamping: 0.6, containerRadius: 55, gravityStrength: 0.04, gravityEnabled: false,
-        interactionRadius: 10.0, attractionStrength: 0.15, repellingStrength: 0.30, vortexStrength: 0.15, stirStrength: 0.10, showInteractionRadius: true,
-        particleBaseColor: '#40E0D0', particleVelocityColorScale: 3.0, bgColor: '#050510',
-        obstacles: [ { x: -25, y: 20, radius: 10 }, { x: 30, y: -15, radius: 14 }, { x: 0, y: -35, radius: 6 }],
-        emitters: [],
-        forceZones: [],
-        mouseTool: 'default',
-    },
-    'High': {
-        particlesCount: 10000,
-        damping: 0.95,
-        maxVelocity: 3.0,
-        particleRepulsionRadius: 4.0,
-        repulsionStrength: 0.30,
-        cohesionRadius: 8.0,
-        cohesionStrength: 0.015,
-        particleSize: 2.0,
-        bloomEnabled: true,
-        bloomStrength: 0.5,
-        bloomRadius: 0.4,
-        bloomThreshold: 0.6,
-        wrapScreenEdges: false,
-        enableObstacles: true,
-        edgeDamping: 0.6, containerRadius: 55, gravityStrength: 0.04, gravityEnabled: false,
-        interactionRadius: 10.0, attractionStrength: 0.15, repellingStrength: 0.30, vortexStrength: 0.15, stirStrength: 0.10, showInteractionRadius: true,
-        particleBaseColor: '#40E0D0', particleVelocityColorScale: 3.0, bgColor: '#050510',
-        obstacles: [ { x: -25, y: 20, radius: 10 }, { x: 30, y: -15, radius: 14 }, { x: 0, y: -35, radius: 6 }],
-        emitters: [],
-        forceZones: [],
-        mouseTool: 'default',
-    },
-    'Extreme': {
-        particlesCount: MAX_PARTICLES,
-        damping: 0.94,
-        maxVelocity: 3.5,
-        particleRepulsionRadius: 4.5,
-        repulsionStrength: 0.35,
-        cohesionRadius: 9.0,
-        cohesionStrength: 0.02,
-        particleSize: 2.2,
-        bloomEnabled: true,
-        bloomStrength: 0.6,
-        bloomRadius: 0.5,
-        bloomThreshold: 0.55,
-        wrapScreenEdges: false,
-        enableObstacles: true,
-        edgeDamping: 0.6, containerRadius: 55, gravityStrength: 0.04, gravityEnabled: false,
-        interactionRadius: 10.0, attractionStrength: 0.15, repellingStrength: 0.30, vortexStrength: 0.15, stirStrength: 0.10, showInteractionRadius: true,
-        particleBaseColor: '#40E0D0', particleVelocityColorScale: 3.0, bgColor: '#050510',
-        obstacles: [ { x: -25, y: 20, radius: 10 }, { x: 30, y: -15, radius: 14 }, { x: 0, y: -35, radius: 6 }],
-        emitters: [],
-        forceZones: [],
-        mouseTool: 'default',
-    },
-    'Fountain': {
-        particlesCount: 500, // Start low, emitter will add more
-        damping: 0.98,
-        maxVelocity: 4.0,
-        particleRepulsionRadius: 1.5,
-        repulsionStrength: 0.1,
-        cohesionRadius: 3.0,
-        cohesionStrength: 0.002,
-        particleSize: 1.2,
-        bloomEnabled: true, bloomStrength: 0.35, bloomRadius: 0.25, bloomThreshold: 0.7,
-        enableObstacles: false,
-        wrapScreenEdges: false,
-        edgeDamping: 0.5, containerRadius: 60, gravityStrength: 0.09, gravityEnabled: true,
-        interactionRadius: 12.0, attractionStrength: 0.2, repellingStrength: 0.35, vortexStrength: 0.2, stirStrength: 0.15, showInteractionRadius: true,
-        particleBaseColor: '#2299FF', particleVelocityColorScale: 2.0, bgColor: '#080818',
-        obstacles: [],
-        emitters: [
-            { id: 'fountain_emitter_1', enabled: true, x: 0, y: -45, rate: 15, // particles per frame
-              initialVelocity: { x: 0, y: 1.8 }, velocityVariance: { x: 0.3, y: 0.2 },
-              color: '#66CCFF', lifespan: 0 } // lifespan 0 = infinite
-        ],
-        forceZones: [],
-        mouseTool: 'default',
-    },
-    'Wind Tunnel': {
-        particlesCount: 1000,
-        damping: 0.99, // Less damping for particles to flow
-        maxVelocity: 5.0,
-        particleRepulsionRadius: 2.0,
-        repulsionStrength: 0.05,
-        cohesionRadius: 4.0,
-        cohesionStrength: 0.001,
-        particleSize: 1.0,
-        bloomEnabled: true, bloomStrength: 0.2, bloomRadius: 0.1, bloomThreshold: 0.8,
-        enableObstacles: true,
-        obstacles: [ { x: 0, y: 0, radius: 12 } ],
-        wrapScreenEdges: false,
-        edgeDamping: 0.1, containerRadius: 60, gravityStrength: 0.0, gravityEnabled: false, // No gravity
-        interactionRadius: 8.0, attractionStrength: 0.1, repellingStrength: 0.2, vortexStrength: 0.1, stirStrength: 0.05, showInteractionRadius: false,
-        particleBaseColor: '#CCCCCC', particleVelocityColorScale: 4.0, bgColor: '#101010',
-        emitters: [
-            { id: 'wind_emitter_1', enabled: true, x: -50, y: 0, rate: 20,
-              initialVelocity: { x: 1.0, y: 0 }, velocityVariance: { x: 0.1, y: 0.5 },
-              color: '#EEEEEE', lifespan: 0 }
-        ],
-        forceZones: [
-            { id: 'wind_zone_1', enabled: true, x: 0, y: 0, radius: 65, force: { x: 0.02, y: 0 } } // Gentle push to the right
-        ],
-        mouseTool: 'moveObstacle',
-    },
-    'Galaxy Spin': {
-        particlesCount: 3000,
-        damping: 0.985,
-        maxVelocity: 3.0,
-        particleRepulsionRadius: 1.0, // Less repulsion for denser core
-        repulsionStrength: 0.02,
-        cohesionRadius: 15.0, // Stronger, wider cohesion
-        cohesionStrength: 0.008,
-        particleSize: 1.3,
-        bloomEnabled: true, bloomStrength: 0.6, bloomRadius: 0.5, bloomThreshold: 0.5,
-        enableObstacles: false,
-        wrapScreenEdges: false,
-        edgeDamping: 0.0, containerRadius: 70, // Larger, no edge damp to keep particles in
-        gravityStrength: 0.0, gravityEnabled: false,
-        interactionRadius: 25.0, attractionStrength: 0.05, repellingStrength: 0.0, vortexStrength: 0.25, stirStrength: 0.0, showInteractionRadius: true,
-        particleBaseColor: '#FFDD88', particleVelocityColorScale: 5.0, bgColor: '#020008',
-        obstacles: [],
-        emitters: [
-            { id: 'galaxy_core_1', enabled: true, x: 0, y: 0, rate: 5,
-              initialVelocity: { x: 0, y: 0 }, velocityVariance: { x: 0.2, y: 0.2 },
-              color: '#FFFFaa', lifespan: 0 }
-        ],
-        forceZones: [],
-        mouseTool: 'default', // Default is fine, vortex is primary interaction here
-    },
-    'Aurora': {
-        particlesCount: 8000,
-        damping: 0.96,
-        maxVelocity: 2.5,
-        particleRepulsionRadius: 3.0,
-        repulsionStrength: 0.2,
-        cohesionRadius: 8.0,
-        cohesionStrength: 0.015,
-        particleSize: 1.8,
-        bloomEnabled: true,
-        bloomStrength: 0.7,
-        bloomRadius: 0.6,
-        bloomThreshold: 0.4,
-        wrapAroundEdges: true,
-        wrapScreenEdges: false,
-        useVerticalGradient: true,
-        gradientHueScale: 0.35,
-        enableObstacles: false,
-        edgeDamping: 0.4, containerRadius: 60, gravityStrength: 0.0, gravityEnabled: false,
-        interactionRadius: 12.0, attractionStrength: 0.15, repellingStrength: 0.25, vortexStrength: 0.1, stirStrength: 0.15, showInteractionRadius: true,
-        particleBaseColor: '#90eeff', particleVelocityColorScale: 4.5, bgColor: '#03120d',
-        obstacles: [],
-        emitters: [
-            { id: 'aurora_top', enabled: true, x: 0, y: 40, rate: 10,
-              initialVelocity: { x: 0, y: -1.5 }, velocityVariance: { x: 0.5, y: 0.5 },
-              color: '#aaffcc', lifespan: 0 },
-            { id: 'aurora_bottom', enabled: true, x: 0, y: -40, rate: 10,
-              initialVelocity: { x: 0, y: 1.5 }, velocityVariance: { x: 0.5, y: 0.5 },
-              color: '#cc88ff', lifespan: 0 }
-        ],
-        forceZones: [],
-        mouseTool: 'default',
-    },
-    'Water Flow': {
-        particlesCount: 6000,
-        damping: 0.97,
-        maxVelocity: 2.0,
-        particleRepulsionRadius: 2.5,
-        repulsionStrength: 0.15,
-        cohesionRadius: 6.0,
-        cohesionStrength: 0.01,
-        particleSize: 2.0,
-        bloomEnabled: false,
-        wrapAroundEdges: true,
-        wrapScreenEdges: true,
-        useVerticalGradient: false,
-        gradientHueScale: 0.0,
-        enableObstacles: false,
-        edgeDamping: 0.3, containerRadius: 60, gravityStrength: 0.0, gravityEnabled: false,
-        interactionRadius: 10.0, attractionStrength: 0.15, repellingStrength: 0.3, vortexStrength: 0.15, stirStrength: 0.1, showInteractionRadius: true,
-        particleBaseColor: '#3399ff', particleVelocityColorScale: 2.5, bgColor: '#050510',
-        obstacles: [],
-        emitters: [],
-        forceZones: [],
-        mouseTool: 'default',
-    },
-};
-
+import { FRUSTUM_SIZE, MAX_PARTICLES, DEFAULT_PRESET_NAME } from './simulationConstants.js';
+import PRESETS from './presets.js';
 
 class FluidSimulation {
     constructor(container) {
