+++ simulations/fluid-simulation/assets/js/simulationConstants.js
@@ -1,0 +1,3 @@
+export const FRUSTUM_SIZE = 120;
+export const MAX_PARTICLES = 20000; // Max particles buffer can hold
+export const DEFAULT_PRESET_NAME = 'Medium';
